More optimisations (I think? Can't recall to be honest) and lots more messing about trying to figure out why the big jitter problem exists -- unsuccessfully.
I even tried loading this version into my ATmega64A (hacekd) transmitter, only
to see the exact same problem, whereas with the /branches/frsky/gruvin9x.hex
firmware, there's no jitter at all -- zero!

What doesn't make any sense is that it's definitely the USART0_RX ISR that
causes (or triggers) the problem -- yet that function is now doing massively
less than it used to be! (The non-jitter firmware processes entire frsky
packets between USART received bytes ffs! This one only loads a single byte
into a buffer then exits. Truly MUCH less work being done. :/)

- - - -


For what it's worth, here's some notes I made while trying to track this down
...

Disabling processFrskyPacket() -- does nothing.

Having frskyGetUserData() always return 0 (bytes) -- does nothing.

Disabling frskyParseRxData() -- does nothing. (Even not calling it at all, does
nothing.)

Disabling frskyTransmitBuffer() -- does nothing.

Disabling FRSKY10mspoll -- does nothing.

................

Having ISR(USART0_RX_vect) not stuff the receive buffer **FIXES** jitter --
actually reduces it to just 1us -- not completely fixes it. But not enabling
the USART receiver at all gives no improvement. (My other unit has zero jitter.
Why not this one?)

So we know that jitter is being caused by the receiver routine -- even merely
stuffing the receive buffer -- only one byte at a time. BUT In the other
unit, we process entire frisky packets between incoming USART bytes without
issue! But this and only this stops the jitter. :/ Eh?!?!?!?!?!

Could it be a difference in interrupt priority between the AT'64 and the
AT'2561? ... NO! Checked that by installing the same code in an AT'64A unit.

+++ branches/frsky/src/gruvin9x.h
@@ -427,8 +427,8 @@ extern uint8_t Timer2_running ;
 extern uint16_t timer2 ;
 void resetTimer2() ;
 
-extern uint16_t g_tmr1Latency_max;
-extern uint16_t g_tmr1Latency_min;
+extern uint8_t g_tmr1Latency_max;
+extern uint8_t g_tmr1Latency_min;
 extern uint16_t g_timeMain;
 extern uint16_t g_time_per10;
